add support for iOS compilation
- use KIVYIOSROOT to see if we are running under the kivy-ios toolchain
- avoid to export the same symbol in all the .so, declaring them as
static resolve the issue
